+++ CHANGELOG.md
@@ -29,7 +29,7 @@
 * [#2436](https://github.com/ruby-grape/grape/pull/2436): Update coverallsapp github-action - [@ericproulx](https://github.com/ericproulx).
 * [#2434](https://github.com/ruby-grape/grape/pull/2434): Implement nested `with` support in parameter dsl - [@numbata](https://github.com/numbata).
 * [#2438](https://github.com/ruby-grape/grape/pull/2438): Fix some Rack::Lint - [@ericproulx](https://github.com/ericproulx).
-* [#2437](https://github.com/ruby-grape/grape/pull/2437): Add Length Validator - [@dhruvCW](https://github.com/dhruvCW).
+* [#2437](https://github.com/ruby-grape/grape/pull/2437): Add length Validator - [@dhruvCW](https://github.com/dhruvCW).
 * Your contribution here.
 
 #### Fixes